requirements.txt
appdirs==1.4.4
blinker==1.9.0
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
contourpy==1.3.1
cycler==0.12.1
dnspython==2.7.0
Flask==3.1.0
fonttools==4.55.0
fpdf==1.7.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==3.0.2
matplotlib==3.9.2
numpy==2.1.3
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pymongo==4.10.1
pyparsing==3.2.0
PyPDF2==3.0.1
python-dateutil==2.9.0.post0
python-nmap==0.7.1
pytz==2024.2
reportlab==4.2.5
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3
vulners==2.2.3
Werkzeug==3.1.3
